refactor(web): give each shared service-card atom its own file and namespace
The settings AI page's shared-ui module had grown five components with
independent consumers, so each moves to its own file under the shared
components tier, and their copy moves from the settings namespace to
common, where cross-domain components read from. ModeToggle stays
private beside ServiceCard, its only consumer.

The domain verification chip also sheds a hand-loosened string status
for the generated DomainVerificationStatusStatusEnum, which surfaced
that an API-returned unknown status rendered as a verification failure;
it now renders the unknown chip, and the never-read message prop is
gone from the chip and its call sites.
