+++ clients/web/src/i18n/locales/es/settings.json
@@ -1803,21 +1803,6 @@
     "warningLabel": "Evento de advertencia",
     "warningSentToast": "Evento de advertencia de Sentry enviado."
   },
-  "sharedUi": {
-    "checkingDomain": "Comprobando dominio…",
-    "domainVerified": "Dominio verificado",
-    "failedTitle": "La verificación del dominio falló. Es posible que los registros DNS no se hayan propagado correctamente. Puedes intentar liberar y volver a registrar el dominio.",
-    "managedLabel": "Administrado",
-    "managedServicesBanner": "Los servicios administrados se miden y se deducen del saldo de tu cuenta de Vellum. <link>Ver precios</link>",
-    "modeToggleAriaLabel": "Modo de servicio",
-    "unknownStatus": "Estado desconocido",
-    "unknownStatusTitle": "No se pudo obtener el estado de verificación del dominio.",
-    "verificationFailed": "Verificación fallida",
-    "verifiedTitle": "Los registros DNS se han verificado. Tu dominio está listo para enviar y recibir correo.",
-    "verifyingDomain": "Verificando dominio…",
-    "verifyingTitle": "Los registros DNS se han aprovisionado. Esperando a que el proveedor de correo los verifique; esto suele tardar unos minutos.",
-    "yourOwnLabel": "Propio"
-  },
   "shortcutsSections": {
     "alreadyUsedBy": "Ya está en uso por {label}",
     "appDescription": "Funcionan mientras una ventana de Vellum está enfocada.",
